# server/swagger/ecmr-api.json
{
  "swagger": "2.0",
  "info": {
    "title": "Digital ECMR APIs",
    "description": "APIs for managing the digital ECMR",
    "version": "1.0.0"
  },
  "host": "localhost:8080",
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "UserSecurity": {
      "type": "basic"
    }
  },
  "basePath": "/api/v1",
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/ECMR/vehicle/vin/{vin}": {
      "get": {
        "x-swagger-router-controller": "ECMRController",
        "operationId": "getAllEcmrsFromVehicleByVin",
        "summary": "Returns the ECMRs containing a vehicle with the provided vin",
        "description": "Returns the ECMRs containing a vehicle with the provided vin",
        "parameters": [
          {
            "name": "x-access-token",
            "in": "header",
            "description": "The authentication token",
            "required": true,
            "type": "string"
          },
          {
            "name": "vin",
            "in": "path",
            "type": "string",
            "description": "The unique vin of a vehicle",
            "required": "true"
          }
        ],
        "responses": {
          "200": {
            "description": "The list contains the ECMRs containing a vehicle with the provided vin, or is empty if they do not exist or if the user has no permission to read them\n",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/ECMR"
              }
            }
          },
          "default": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/ECMR/vehicle/plateNumber/{plateNumber}": {
      "get": {
        "x-swagger-router-controller": "ECMRController",
        "operationId": "getAllEcmrsFromVehicleByPlateNumber",
        "summary": "Returns the ECMRs containing a vehicle with the provided platenumber",
        "description": "Returns the ECMRs containing a vehicle with the provided platenumber",
        "parameters": [
          {
            "name": "x-access-token",
            "in": "header",
            "description": "The authentication token",
            "required": true,
            "type": "string"
          },
          {
            "name": "plateNumber",
            "in": "path",
            "type": "string",
            "description": "The unique plateNumber of a vehicle",
            "required": "true"
          }
        ],
        "responses": {
          "200": {
            "description": "The list contains the ECMRs containing a vehicle with the provided platenumber, or is empty if they do not exist or if the user has no permission to read them\n",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/ECMR"
              }
            }
          },
          "default": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
...
    "/vehicle/": {
      "post": {
        "x-swagger-router-controller": "VehicleController",
        "operationId": "createVehicles",
        "summary": "Creates a set of vehicles from a list",
        "description": "Creates a set of vehicles from a list, if the user has the proper permissions.\n",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "x-access-token",
            "in": "header",
            "description": "The authentication token",
            "required": true,
            "type": "string"
          },
          {
            "name": "vehicles",
            "in": "body",
            "description": "The list of vehicles",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Vehicle"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The vehicles have been created and written into the blockchain\n",
            "type": "array",
            "items": {
              "$ref": "#/definitions/Vehicle"
            }
          },
          "default": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      },
      "get": {
        "x-swagger-router-controller": "VehicleController",
        "operationId": "getAllVehicles",
        "summary": "Returns the complete list of vehicles",
        "description": "Returns the complete list of all vehicles which the current user has the right to read",
        "parameters": [
          {
            "name": "x-access-token",
            "in": "header",
            "description": "The authentication token",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "The vehicles list is returned\n",
            "type": "array",
            "items": {
              "$ref": "#/definitions/Vehicle"
            }
          },
          "default": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/vehicle/vin/{vin}": {
      "get": {
        "x-swagger-router-controller": "VehicleController",
        "operationId": "getVehicleByVin",
        "summary": "Returns the vehicle having the provided vin",
        "description": "Returns the vehicle having the provided vin",
        "parameters": [
          {
            "name": "x-access-token",
            "in": "header",
            "description": "The authentication token",
            "required": true,
            "type": "string"
          },
          {
            "name": "vin",
            "in": "path",
            "description": "The vin of the vehicle",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "The vehicles is returned\n",
            "type": "array",
            "items": {
              "$ref": "#/definitions/Vehicle"
            }
          },
          "default": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/vehicle/plateNumber/{plateNumber}": {
      "get": {
        "x-swagger-router-controller": "VehicleController",
        "operationId": "getVehicleByPlateNumber",
        "summary": "Returns the vehicle having the provided plateNumber",
        "description": "Returns the vehicle having the provided plateNumber",
        "parameters": [
          {
            "name": "x-access-token",
            "in": "header",
            "description": "The authentication token",
            "required": true,
            "type": "string"
          },
          {
            "name": "plateNumber",
            "in": "path",
            "description": "The plateNumber of the vehicle",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "The vehicles is returned\n",
            "type": "array",
            "items": {
              "$ref": "#/definitions/Vehicle"
            }
          },
          "default": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "Address": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "required": true
        },
        "street": {
          "type": "string",
          "required": true
        },
        "houseNumber": {
          "type": "string",
          "required": true
        },
        "city": {
          "type": "string",
          "required": true
        },
        "zipCode": {
          "type": "string",
          "required": true
        },
        "country": {
          "type": "string",
          "required": true
        },
        "latitude": {
          "type": "number",
          "required": true
        },
        "longitude": {
          "type": "number",
          "required": true
        }
      }
    },
    "Delivery": {
      "type": "object",
      "properties": {
        "Address": {
          "$ref": "#/definitions/Address",
          "required": true
        },
        "date": {
          "type": "number",
          "required": false
        }
      }
    },
    "Loading": {
      "type": "object",
      "properties": {
        "Address": {
          "$ref": "#/definitions/Address",
          "required": true
        },
        "date": {
          "type": "number",
          "required": true
        }
      }
    },
    "Creation": {
      "type": "object",
      "properties": {
        "Address": {
          "$ref": "#/definitions/Address",
          "required": true
        },
        "date": {
          "type": "number",
          "required": true
        }
      }
    },
    "Remark": {
      "type": "object",
      "properties": {
        "comments": {
          "type": "string",
          "required": true
        },
        "isDamaged": {
          "type": "boolean",
          "required": false
        }
      }
    },
    "Vehicle": {
      "type": "object",
      "properties": {
        "vin": {
          "type": "string",
          "required": true
        },
        "manufacturer": {
          "type": "boolean",
          "required": true
        },
        "model": {
          "type": "string",
          "required": true
        },
        "type": {
          "type": "boolean",
          "required": true
        },
        "emcrs": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "required": true
        },
        "odoMeterReading": {
          "type": "number",
          "required": true
        },
        "plateNumber": {
          "type": "string",
          "required": true
        },
        "registrationCountry": {
          "type": "string",
          "required": false
        }
      }
    },
    "ECMR_STATUS": {
      "type": "string",
      "enum": [
        "CREATED",
        "LOADED",
        "IN_TRANSIT",
        "DELIVERED",
        "CONFIRMED_DELIVERED"
      ]
    },
    "ECMR": {
      "type": "object",
      "description": "description",
      "properties": {
        "ecmrID": {
          "type": "string",
          "required": true
        },
        "status": {
          "$ref": "#/definitions/ECMR_STATUS",
          "required": true
        },
        "creation": {
          "$ref": "#/definitions/Creation",
          "required": true
        },
        "loading": {
          "$ref": "#/definitions/Loading",
          "required": true
        },
        "delivery": {
          "$ref": "#/definitions/Delivery",
          "required": true
        },
        "owner": {
          "type": "string",
          "description": "The id of the legal owner org",
          "required": true
        },
        "source": {
          "type": "string",
          "description": "The id of the compound org",
          "required": true
        },
        "transporter": {
          "type": "string",
          "description": "The id of the transporter user",
          "required": false
        },
        "carrier": {
          "type": "string",
          "description": "The id of the carrier org",
          "required": true
        },
        "recipientOrg": {
          "type": "string",
          "description": "The id of the recipient org",
          "required": true
        },
        "recipient": {
          "type": "string",
          "description": "The id of the recipient user",
          "required": false
        },
        "issueDate": {
          "type": "number",
          "required": true
        },
        "issuedBy": {
          "type": "string",
          "description": "The id of the Legal Owner org who issued the ECMR",
          "required": true
        },
        "carrierComments": {
          "type": "string",
          "required": true
        },
        "goods": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Good"
          },
          "required": true
        },
        "legalOwnerInstructions": {
          "type": "string",
          "required": true
        },
        "paymentInstructions": {
          "type": "string",
          "required": true
        },
        "payOnDelivery": {
          "type": "string",
          "required": true
        },
        "documents": {
          "type": "string",
          "required": false
        },
        "compoundSignature": {
          "$ref": "#/definitions/Signature",
          "required": false
        },
        "carrierLoadingSignature": {
          "$ref": "#/definitions/Signature",
          "required": false
        },
        "carrierDeliverySignature": {
          "$ref": "#/definitions/Signature",
          "required": false
        },
        "recipientSignature": {
          "$ref": "#/definitions/Signature",
          "required": false
        }
      }
    },
...
    "Good": {
      "type": "object",
      "properties": {
        "vehicle": {
          "$ref": "#/definitions/Vehicle",
          "required": true
        },
        "carrierLoadingRemark": {
          "$ref": "#/definitions/Remark",
          "required": false
        },
        "compoundRemark": {
          "$ref": "#/definitions/Remark",
          "required": false
        },
        "recipientRemark": {
          "$ref": "#/definitions/Remark",
          "required": false
        },
        "carrierDeliveryRemark": {
          "$ref": "#/definitions/Remark",
          "required": false
        },
        "description": {
          "type": "string",
          "required": false
        },
        "weight": {
          "type": "number",
          "required": false
        }
      }
    },
...
    "Signature": {
      "type": "object",
      "properties": {
        "certificate": {
          "type": "string",
          "required": true
        },
        "latitude": {
          "type": "number",
          "required": false
        },
        "longitude": {
          "type": "string",
          "required": true
        },
        "ip": {
          "type": "string",
          "required": false
        },
        "timestamp": {
          "type": "number",
          "required": true
        },
        "generalRemark": {
          "$ref": "#/definitions/Remark",
          "description": "General remarks for the signature",
          "required": true
        }
      }
    },
...
    "Error": {
      "type": "object",
      "properties": {
        "description": {
          "type": "string"
        },
        "message": {
          "type": "string"
        }
      }
    }
  }
}